Fixed long to int conversion in MitraHint

+++ AnodyneArchipelago/MitraHint.cs
@@ -5,9 +5,9 @@ namespace AnodyneArchipelago
     public readonly struct MitraHint
     {
         [JsonProperty("item")]
-        public readonly int itemID;
+        public readonly long itemID;
         [JsonProperty("location")]
-        public readonly int locationID;
+        public readonly long locationID;
         [JsonProperty("location_player")]
         public readonly int playerSlot;
     }
